src/test/java/dev/kauanmocelin/bank/account/adapter/out/persistence/AccountPersistenceAdapterTest.java
package dev.kauanmocelin.bank.account.adapter.out.persistence;

import dev.kauanmocelin.bank.account.domain.Account;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.context.annotation.Import;
import org.springframework.test.context.jdbc.Sql;

import static org.assertj.core.api.Assertions.assertThat;

@DataJpaTest
@Import({AccountPersistenceAdapter.class, AccountMapper.class})
class AccountPersistenceAdapterTest {

    @Autowired
    private AccountPersistenceAdapter accountPersistenceAdapter;

    @Test
    @Sql("AccountPersistenceAdapterTest.sql")
    void loadsAccount() {
        Account account = accountPersistenceAdapter.loadAccount(1L);

        assertThat(account.getBalance()).isEqualTo(100.00);
    }
}